New Fireball/Flame Effect Shaders & VFX Graphs
Updates to fireball and smoke shaders, new Fireball VFX progress, may require some systems tweaking to orient to direction, currently oriented to camera, and velocity (which has been set to an arbitrary tiny value in a direction), but will need to update this according to direction of prefab travel. Lots of other experimentation with trails, worldspace particles etc.